fix: moving over the notebook specific pipe
