+++ ui/src/notebooks/components/PipeList.tsx
@@ -1,34 +1,6 @@
-import React, {FC, useContext, useCallback, createElement, useMemo} from 'react'
-import {PipeContextProps, PipeData} from 'src/notebooks'
-import Pipe from 'src/notebooks/components/Pipe'
+import React, {FC, useContext, useCallback} from 'react'
 import {NotebookContext} from 'src/notebooks/context/notebook'
-import NotebookPanel from 'src/notebooks/components/panel/NotebookPanel'
-
-interface NotebookPipeProps {
-  index: number
-  data: PipeData
-  onUpdate: (index: number, pipe: PipeData) => void
-}
-
-const NotebookPipe: FC<NotebookPipeProps> = ({index, data, onUpdate}) => {
-  const panel: FC<PipeContextProps> = useMemo(
-    () => props => {
-      const _props = {
-        ...props,
-        index,
-      }
-
-      return createElement(NotebookPanel, _props)
-    },
-    [index]
-  )
-
-  const _onUpdate = (data: PipeData) => {
-    onUpdate(index, data)
-  }
-
-  return <Pipe data={data} onUpdate={_onUpdate} Context={panel} />
-}
+import NotebookPipe from 'src/notebooks/components/NotebookPipe'
 
 const PipeList: FC = () => {
   const {id, pipes, updatePipe} = useContext(NotebookContext)
